serve client works on diffdock and stores directory outputs
