+++ ganglia/client.py
@@ -1,11 +1,14 @@
+import asyncio
 import json
+import websockets.exceptions
+import websockets.client as wsc
 
 
 def generate_diffdock_instructions(
     debug_logs=True,
-    protein="1a30/1a30_protein.pdb",
-    ligand="1a30/1a30_ligand.sdf",
-    output="1a30", 
+    protein="6jap/6jap_protein_processed.pdb",
+    ligand="6jap/6jap_ligand.sdf",
+    output="6jap", 
     repr_layers=33,
     inference_steps=20,
     samples_per_complex=40,
@@ -19,7 +22,7 @@ def generate_diffdock_instructions(
         "long_args": {"gpus": "all"},
         "cmd": (
             '/bin/bash -c \"'
-            f'mkdir -p /outputs/{output}&& ' 
+            f'mkdir -p /outputs/{output} && ' 
             f'python datasets/esm_embedding_preparation.py --protein_path /inputs/{protein} --out_file /outputs/{output}/esm2_input.fasta '
             f'&& HOME=esm/model_weights python esm/scripts/extract.py esm2_t33_650M_UR50D /outputs/{output}/esm2_input.fasta '
             f'/outputs/{output}/esm2_output --repr_layers {repr_layers} --include per_tok && '
@@ -76,5 +79,18 @@ def generate_gnina_instructions(
     }
     return json.dumps(instructions)
 
-if __name__ == "__main__":
-    print(generate_vina_instructions())
+async def run_with_socket(instructions) -> None:
+    host = "localhost"
+    uri = f"ws://{host}:8765"
+
+    # TODO improve server loop so ping_timeout can get set to normal val
+    async with wsc.connect(uri, ping_timeout=None) as ws:
+        await ws.send(instructions)
+        try:
+            while True:
+                result = await ws.recv()
+                print(f'From socket: {result}', end="")  # `end=''` removes extra newline
+        except ConnectionError as ce:
+            print(f"Error: {ce}")
+        except websockets.exceptions.ConnectionClosedOK as cco:
+            print(f"Success: {cco}")
