+++ server/helpers/settings.py
@@ -1,0 +1,5 @@
+import pathlib
+
+
+# path of server.py
+PROJECT_ROOT = pathlib.Path(__file__).parent.parent.resolve()
